+++ tests/test_connect.py
@@ -4,6 +4,8 @@ from mock import patch
 
 from src import errors
 from src.client import Client
+from src.protocol import ClientPacketTypes, ServerPacketTypes
+from src.reader import _read_one
 from tests.testcase import BaseTestCase
 
 
@@ -20,6 +22,13 @@ class PacketsTestCase(BaseTestCase):
         exc = e.exception
         self.assertIn('Code:', str(exc))
         self.assertIn('Stack trace:', str(exc))
+
+    def test_packets_to_str(self):
+        self.assertEqual(ClientPacketTypes.to_str(2), 'Data')
+        self.assertEqual(ClientPacketTypes.to_str(42), 'Unknown packet')
+
+        self.assertEqual(ServerPacketTypes.to_str(4), 'Pong')
+        self.assertEqual(ClientPacketTypes.to_str(42), 'Unknown packet')
 
 
 class SocketErrorTestCase(BaseTestCase):
@@ -58,3 +67,44 @@ class SocketErrorTestCase(BaseTestCase):
                 # New socket should be created.
                 rv = self.client.execute('SELECT 1')
                 self.assertEqual(rv, [(1, )])
+
+    def test_socket_error_on_ping(self):
+        self.client.execute('SELECT 1')
+
+        def side_effect(*args, **kwargs):
+            raise socket.error(32, 'Broken pipe')
+
+        with patch.object(self.client.connection, 'fout') as mocked_fout:
+            mocked_fout.flush.side_effect = side_effect
+
+            rv = self.client.execute('SELECT 1')
+            self.assertEqual(rv, [(1, )])
+
+    def test_ping_got_unexpected_package(self):
+        self.client.execute('SELECT 1')
+
+        with patch.object(self.client.connection, 'fin') as mocked_fin:
+            # Emulate Exception packet on ping.
+            mocked_fin.read.return_value = b'\x02'
+
+            with self.assertRaises(errors.UnexpectedPacketFromServerError):
+                self.client.execute('SELECT 1')
+
+    def test_eof_error_on_ping(self):
+        self.client.execute('SELECT 1')
+
+        self.raised = False
+
+        def side_effect(*args, **kwargs):
+            if not self.raised:
+                self.raised = True
+                raise EOFError('Unexpected EOF while reading bytes')
+
+            else:
+                return _read_one(*args, **kwargs)
+
+        with patch('src.reader._read_one') as mocked_fin:
+            mocked_fin.side_effect = side_effect
+
+            rv = self.client.execute('SELECT 1')
+            self.assertEqual(rv, [(1, )])